+++ apps/weather_demo.cpp
@@ -14,7 +14,7 @@ int main() {
     // Initialize weather system
     WeatherSystem weather;
     Lightning lightning;
-    
+
     try {
         // Try to load weather config
         weather.loadFromYaml("config/weather.yaml");
@@ -23,99 +23,99 @@ int main() {
         printf("⚠ Could not load weather config: %s\n", e.what());
         printf("  Using default weather settings\n");
     }
-    
+
     // Simulate weather system over time
     printf("\n--- Weather Simulation ---\n");
-    
+
     const float dt = 0.016f; // 60 FPS
     const int frames = 300;  // 5 seconds
-    
+
     // Test different weather states
     WeatherState states[] = {
         WeatherState::CLEAR,
-        WeatherState::CLOUDY, 
+        WeatherState::CLOUDY,
         WeatherState::RAIN,
         WeatherState::SNOW,
         WeatherState::STORM,
         WeatherState::FOG
     };
-    
+
     const char* stateNames[] = {
         "Clear", "Cloudy", "Rain", "Snow", "Storm", "Fog"
     };
-    
+
     for (int s = 0; s < 6; s++) {
         weather.setState(states[s]);
         printf("\n%s Weather:\n", stateNames[s]);
-        
+
         for (int i = 0; i < frames / 6; i++) {
-            weather.tick(dt);
+            weather.tick(static_cast<double>(dt));
             lightning.tick(dt, states[s] == WeatherState::STORM);
-            
+
             if (i % 10 == 0) { // Print every 10 frames
                 auto ubo = weather.getUBO();
-                printf("  Frame %3d: Wind=%.1f m/s, Clouds=%.2f, Precip=%.1f mm/h", 
-                       i, ubo.windSpeed, ubo.cloudCoverage, ubo.precipRate);
-                
+                printf("  Frame %3d: Wind=%.1f m/s, Clouds=%.2f, Precip=%.1f mm/h",
+                       i, static_cast<double>(ubo.windSpeed), static_cast<double>(ubo.cloudCoverage), static_cast<double>(ubo.precipRate));
+
                 if (lightning.intensity() > 0.1f) {
-                    printf(" ⚡Lightning: %.2f", lightning.intensity());
+                    printf(" ⚡Lightning: %.2f", static_cast<double>(lightning.intensity()));
                 }
                 printf("\n");
             }
         }
     }
-    
+
     printf("\n--- Frame Graph Demo ---\n");
-    
+
     // Demonstrate frame graph usage
     FrameGraphMin fg;
     FrameCtx ctx{};
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 1: Update Weather UBO\n"); 
-        weather.tick(dt);
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 1: Update Weather UBO\n");
+        weather.tick(static_cast<double>(dt));
         auto ubo = weather.getUBO();
         c.user = &ubo; // Store UBO for next passes
     });
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 2: Sky Fullscreen Render\n"); 
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 2: Sky Fullscreen Render\n");
         // Would render sky with Hosek-Preetham model
     });
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 3: Clouds Fullscreen Overlay\n"); 
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 3: Clouds Fullscreen Overlay\n");
         // Would render clouds with fbm noise
     });
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 4: Opaque Geometry\n"); 
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 4: Opaque Geometry\n");
         // Would render world geometry with weather material modifiers
     });
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 5: Precipitation Update + Render\n"); 
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 5: Precipitation Update + Render\n");
         // Would dispatch compute for particle update, then render sprites
     });
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 6: Temporal Accumulation (Clouds)\n"); 
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 6: Temporal Accumulation (Clouds)\n");
         // Would apply temporal reprojection to stabilize clouds
     });
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 7: Temporal Accumulation (Precipitation)\n"); 
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 7: Temporal Accumulation (Precipitation)\n");
         // Would apply temporal reprojection to stabilize precipitation
     });
-    
-    fg.add([&](FrameCtx& c) { 
-        printf("Pass 8: Height Fog\n"); 
+
+    fg.add([&](FrameCtx& c) {
+        printf("Pass 8: Height Fog\n");
         // Would apply height fog as final post process
     });
-    
+
     printf("Executing frame graph:\n");
     fg.execute(ctx);
-    
+
     printf("\n--- Console Commands Demo ---\n");
     printf("Example weather console commands:\n");
     printf("  wx.set STORM              # Set weather to storm\n");
@@ -125,24 +125,24 @@ int main() {
     printf("  wx.fog 0.02               # Set fog density\n");
     printf("  wx.sun.az 135             # Set sun azimuth\n");
     printf("  wx.lightning on           # Enable lightning\n");
-    
+
     printf("\n--- Weather System Stats ---\n");
     auto finalUbo = weather.getUBO();
     printf("Final Weather State:\n");
-    printf("  Wind Direction: (%.2f, %.2f, %.2f)\n", 
-           finalUbo.windDir.x, finalUbo.windDir.y, finalUbo.windDir.z);
-    printf("  Wind Speed: %.2f m/s\n", finalUbo.windSpeed);
-    printf("  Cloud Coverage: %.2f\n", finalUbo.cloudCoverage);
-    printf("  Turbidity: %.2f\n", finalUbo.turbidity);
-    printf("  Time of Day: %.2f\n", finalUbo.timeOfDay);
+    printf("  Wind Direction: (%.2f, %.2f, %.2f)\n",
+           static_cast<double>(finalUbo.windDir.x), static_cast<double>(finalUbo.windDir.y), static_cast<double>(finalUbo.windDir.z));
+    printf("  Wind Speed: %.2f m/s\n", static_cast<double>(finalUbo.windSpeed));
+    printf("  Cloud Coverage: %.2f\n", static_cast<double>(finalUbo.cloudCoverage));
+    printf("  Turbidity: %.2f\n", static_cast<double>(finalUbo.turbidity));
+    printf("  Time of Day: %.2f\n", static_cast<double>(finalUbo.timeOfDay));
     printf("  Sun Direction: (%.2f, %.2f, %.2f)\n",
-           finalUbo.sunDir.x, finalUbo.sunDir.y, finalUbo.sunDir.z);
-    printf("  Fog Density: %.4f\n", finalUbo.fogDensity);
-    printf("  Lightning Intensity: %.2f\n", finalUbo.lightning);
-    
+           static_cast<double>(finalUbo.sunDir.x), static_cast<double>(finalUbo.sunDir.y), static_cast<double>(finalUbo.sunDir.z));
+    printf("  Fog Density: %.4f\n", static_cast<double>(finalUbo.fogDensity));
+    printf("  Lightning Intensity: %.2f\n", static_cast<double>(finalUbo.lightning));
+
     printf("\n✓ Weather system demo completed successfully!\n");
     printf("  The weather system is now integrated and ready for use.\n");
     printf("  All shaders have been created and the frame graph is wired.\n");
-    
+
     return 0;
 }
